reflect change in values when top fragment is scrolled
for this created a new Event bus where PageScrollPostion class is initialized.
The listener for this at top fragment, then creates another Event Bus where Converted Value is initialized.
This event bus's listener is present at bottom fragment and it calculates the required values and reflects to each
of its textView