bug fix : resolve issue where app was crashing due to index out of bound on page scroll after certain limit
Cause : the app uses ScreenSlideTopFragment and ScreenSlideBotFragment for various options provided in the right spinner
As various fragments are rendered when different options in the spinner are selected, certain sets of fragments are created on each spinner's choice
When onPageSelected listener creates an Event.post, the listeners at each of these created fragment listens and tries to complete the task.
However, we only want the fragments from corresponding selectedConversion i.e. option at spinner to change.
But, not all selectedConversions have same dataset and dataset length.
So, when a selectedConversion with higher dataset length, at a higher dataset postion sends an Event.post,
all those previously created fragments are listening.
Some may have lower dataset length, so at that instance they are trying to access data from an index that is not there.
Hence the dataout of bound sort of error.

Solution : Inorder to solve this, passed selectedConversion attribute through PageScrollPosition object as well as
to ScreenSlideTopFragment from its adapter at the time of its creation.

When the listener at ScreenSlideTopFragment receives an Event.post, it compares these two values i.e.
the selectedConversion of the PageScrollPosition and the current Fragment's selectedConversion.

If these two matches, then only it will post an Event bus for changing the values across BottomFragment and TopFragment if needed.
This has led to improvement from the lag experienced earlier as unnecessary values were being updated for
fragments of other selectedConversion.