add abstract class calculateConstant to ProportionalUnit class and change functionality of getConstant method
getConstant method is modified such that any unit that needs to be converted from one unit to other,
is first converted to a reference unit.
(The reference unit for each Unit Conversion model such as Length, Temperature, Weight, etc are set via constructor.)
After being converted to reference unit, it is converted to the final unit in which the output is required.
Inorder to maintain precision and to avoid unwanted calculations, if the to and from units are same,
the constant value is directly returned as 1.